Refactored code and split functions into helper functions
