+++ src/helpers/utils.js
@@ -1,10 +1,22 @@
 import fs from "fs";
 import { URL } from "url";
+import { createDBs, queryTable, updateTable } from "./db.js";
+import { sendJobRun } from "./chainlink.js";
+
+// get environment variables
+const { SEND_CHECK_INTERVAL = "45" } = process.env;
+
+/**
+ * Environment variables validation
+ */
+if (process.env.NODE_ENV != "test") {
+  assert(Number(SEND_CHECK_INTERVAL), "$SEND_CHECK_INTERVAL is required");
+}
 
 /**
  * Function to read a json file
- * @param {*} filename  file name or path to read
- * @returns the JSON data in the file
+ * @param {string} filename  file name or path to read
+ * @returns {Object} the JSON data in the file
  */
 export const readJSONFile = (filename) => {
   let rawdata = fs.readFileSync(filename);
@@ -14,8 +26,8 @@ export const readJSONFile = (filename) => {
 
 /**
  * Function to save JSON data to a file
- * @param {*} newData new JSON data to save
- * @param {*} filename filename to save data to
+ * @param {Object} newData new JSON data to save
+ * @param {string} filename filename to save data to
  */
 export const saveJSONDataToFile = (newData, filename) => {
   let data = JSON.stringify(newData);
@@ -24,8 +36,8 @@ export const saveJSONDataToFile = (newData, filename) => {
 
 /**
  * Function to check whether a URL is valid or not
- * @param {*} url the URL to check
- * @returns whether the url is valid or not
+ * @param {string} url the URL to check
+ * @returns {boolean} whether the url is valid or not
  */
 export const validUrl = (url) => {
   try {
@@ -38,23 +50,53 @@ export const validUrl = (url) => {
 
 /**
  * Function to create a delay
- * @param {*} ms milliseconds to delay
- * @returns a Promise to delay
+ * @param {number} ms milliseconds to delay
+ * @returns {Promise} a Promise to delay
  */
 export const delay = async (ms) => {
   return new Promise(async (res) => await setTimeout(res, ms));
 };
 
 /**
- * Function to read from vstorage
- * @param {*} vstorage vstorage
- * @param {*} feed the feed to read
- * @param {*} roundData whether to read round data or price data
- * @returns a Promise with CapData of result
+ * Function to initialise state
  */
-export const readVStorage = async (vstorage, feed, roundData) => {
-  let key = roundData
-    ? "published.priceFeed." + feed + "_price_feed.latestRound"
-    : "published.priceFeed." + feed + "_price_feed";
-  return await vstorage.readLatest(key);
+export const initialiseState = async () => {
+  //create tables if they do not exist
+  await createDBs();
+};
+
+/**
+ * Function to submit a new job run to the Chainlink node
+ * @param {string} feed the feed to submit a job for
+ * @param {number} requestType the request type to send as a parameter with the job request. 1 if a timer request, 2 if triggered by a price deviation, 3 new round.
+ */
+export const submitNewJob = async (feed, requestType) => {
+  //get latest request id
+  let query = await queryTable("jobs", ["request_id", "id"], feed);
+  let newRequestId = query.request_id + 1;
+  //update table
+  await updateTable(
+    "jobs",
+    { request_id: newRequestId, last_request_sent: Date.now() / 1000 },
+    feed
+  );
+
+  console.log("Sending job spec", feed, "request", newRequestId);
+
+  //send job run
+  await sendJobRun(newRequestId, query.id, requestType);
+};
+
+/**
+ * Function to check if currently in submission
+ * @param {string} feed feed to check for
+ * @returns {boolean} whether last submission was made in less than SEND_CHECK_INTERVAL seconds
+ */
+export const checkIfInSubmission = async (feed) => {
+  //get last submission time
+  let query = await queryTable("jobs", ["last_submission_time"], feed);
+  //get seconds since last price submission
+  let timePassedSinceSubmission =
+    Date.now() / 1000 - query.last_submission_time;
+  return timePassedSinceSubmission < Number(SEND_CHECK_INTERVAL);
 };
